requirements.txt
Django~=4.1.7
djangorestframework~=3.14.0
pillow~=9.4.0
